Load user credentials from database
Query the account and account_password tables in the Permanent database
to fetch the needed information.

Because of how the schema is laid out, it is possible for a user to have
multiple passwords set. The application expects there to only ever be
one row in `account_password` for each `accountId`, but due to previous
bugs, there are a handful of users in production who have more than one.
The application relies on MySQL's implicit ordering when there are
multiple rows; in theory, the order is arbitrary and could be different
from query to query, but from what I've observed tends to be sorted by
primary key. Explicitly opt into that behavior with a correlated
subquery.

We'll reach out to the users affected to make sure they're still able to
log in, although I think if we imported the wrong password they'd be
able to use the "forgot password" functionality without needing to talk
to support.

Note that MySQL encodes boolean values as either 0 or 1, and so some of
the properties returned are of type `number` rather than `boolean`.