Transform Permanent users to Auth0 users
Format the data we get back from the database into the form expected by
Auth0[1].

As noted in the README, we need to convert the PHP-variant bcrypt hashes
into the format Auth0 expects. We also need to format the parsed Date
back into something Auth0 can handle.

The MFA factor logic is slightly tricky. The goal is to only include an
MFA factor if we've already confirmed that it's valid - so not just that
we have an email or a phone number, but that the user has verified that
method of contact. If they haven't, then we don't want a partially
filled out factor, or one with a null value. Do that with array
spreading[3].

[1] https://auth0.com/docs/users/import-and-export-users/bulk-user-import-database-schema-and-examples
[2] https://stackoverflow.com/a/36225192
[3] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/Spread_syntax