Validate and format phone numbers
The Permanent application allows users to enter their phone number, and
then subsequently allows them to validate it and use it as a second
factor while logging in. Twilio appears to be somewhat unconcerned with
the format of such phone numbers, as there are a variety of different
formats in production which have been marked as verified.

Auth0, however, wants phone numbers in the E.164 format[1], which
includes the + prefix and the country code.

To do so, use libphonenumber-js[1] to validate, parse, and format phone
numbers.

[1] https://en.wikipedia.org/wiki/E.164
[1] https://www.npmjs.com/package/libphonenumber-js