Add migrate command to load data into Auth0
Finish up the ETL script with the load command: take the transformed
data and send it off to Auth0.

Use Axios as our http client, with a generous amount of logging. The
Auth0 bulk user import documentation[1] briefly and indirectly specifies
that the request should be a POST with form data, so use form-data[2] to
do so, rather than the manual replacement suggested in its nodejs
example code. After sending the request, poll the job status endpoint,
print the results, and print the errors if the import job is
unsuccessful.

This isn't great code. I took some shortcuts. However, it's only going
to be run once in production, and then retired forever, so... as long as
it works correctly, I'm okay with it being kind of ugly.

[1] https://auth0.com/docs/users/import-and-export-users/bulk-user-imports
[2] https://www.npmjs.com/package/form-data