Skip mfa_factors if empty list
Attempting to import a user who has an empty MFA list - that is, someone
who does not have a verified email and does not have a verified phone
number - causes Auth0 to reject that user. The import job still
completes, just with some failures.

Instead, simply do not include the mfa_factors key if it would only have
an empty list. Extract a function to make that check a little bit more
readable.